Moved transcripts_to_genes to its own file
